Add support for arrays of all supported data types

+++ tests/util/index_helpers.hpp
@@ -21,7 +21,7 @@
     std::initializer_list<size_t> expected = {__VA_ARGS__}; \
     auto actual = index_set.as_indexes(); \
     INFO("Checking " #index_set); \
-    REQUIRE(expected.size() == std::distance(actual.begin(), actual.end())); \
+    REQUIRE(expected.size() == static_cast<size_t>(std::distance(actual.begin(), actual.end()))); \
     auto begin = actual.begin(); \
     for (auto index : expected) { \
         REQUIRE(*begin++ == index); \
